msg/async: send keepalive on connection race winner
If we win a connection race, we could still hit a fault before we finish
connecting, and then go into standby. Ensure that we fully connect so that
we can send our keepalive and the other end is able to send their whatever
to us.

Master PR: https://github.com/ceph/ceph/pull/25754

This does not follow the normal backport process because the protocol of
transport layer of Ceph has been greatly changed for Nautilus.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>