os/filestore: make note of EIO errors when we see them
This is imprecise, since we can't (easily) map an EIO back to a specific
part of the device, or even (easily) tell whether it was a read or write
error.  It's enough to mark a crash dump as an EIO event, though, and to
include the name of the (primary) filestore device.

Signed-off-by: Sage Weil <sage@redhat.com>