mon/OSDMonitor: osd add-no{up,down,in,out} - remove state checker
The existing sanity checkers are total unnecessary, e.g, noup indicates
we should not allow the corresponding osd to boot (marked as up)
later on, which has nothing to do with the current existing osd states.

By removing these checkers, we can:
$ ceph osd add-noup osd.nnn
$ ceph osd down osd.nnn

this way we can manually mark a specific osd as down and then ensure it
will not be automatically marked as up until the specific restrictions
has been resolved.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>