mgr: drop GIL while waiting for mon to set config
This is awkward because we have some set_config users that hold the GIL
and some that do not.  For those that do, we need to drop the GIL while
taking local locks, but hold it while using the PyFormatter, and then
drop it again while doing our blocking request.

Signed-off-by: Sage Weil <sage@redhat.com>