mgr/zabbix: Fix raw_bytes_used key name
This patch fixes raw_bytes_used key which was renamed to stored_raw.
In order not to fail with key errors in case of renamed keys, direct request
for a key has been replaced with get() method. So in case of missing key
just single stat will be missing.
Also added key percent_used and fixed zabbix template to be fully
compatible with zabbix 3.0

Fixes: https://tracker.ceph.com/issues/39644
Signed-off-by: Dmitriy Rabotjagov <noonedeadpunk@ya.ru>