osd: track some heartbeat stats for peer OSDs
We need to keep track of peer epochs and when they last sent or replied
to pings.  Keep this info in a separate ref-counted struct so that we can
point to it directly from other parts of the OSD that will need it.
Eliminate HeartbeatSession and merge the fields into Session since that
is uniformly attached to all incoming connections, and we want to uniformly
handle MOSDPing messages on both incoming and outgoing connections.

Include the ability for a caller to put a PG on a notify list in the
event that a ping or ping reply does arrive.

For now we do nothing with the affected PGs; that will come later.

Signed-off-by: Sage Weil <sage@inktank.com>