client: _readdir_cache_cb() may use the readdir_cache already clear
https://github.com/ceph/ceph/pull/29526
Calling function _readdir_cache_cb() means to get dir information from the cache, but in the while loop,
the contents of readdir_cache are not static.

For example, _getattr() will trigger insert_trace to update readdir_cache, such as:
1. insert_trace()->clear_dir_complete_and_ordered() or
2. insert_trace()->add_update_inode()->add_update_cap()-> check_cap_issue()->clear_dir_complete_and_ordered()
So, there is a possibility that dn is empty after the _getattr() call.

Fixes: http://tracker.ceph.com/issues/41148
Change-Id: I48c60e2156bb4815e1228885b4528fc9320f520e
Signed-off-by: huanwen ren <ren.huanwen@zte.com.cn>