osd/PeeringState: send infos via message_map (not info_map)
Queue up separate message for each.

Note that even though we are currenting sending this via a pg_notify_t,
not all of those fields are needed or consumed on the other end (notably,
PastIntervals is not used by handle_fast_pg_info).  Only pass the fields
we need to send_info() so that we can improve the transport to be less
weird later.

Signed-off-by: Sage Weil <sage@redhat.com>