mon: consolidate 'sync force' commands
Replace the 'ceph [mon] sync force' commands and just use the asok
sync_force command instead.  This is a low-level command that nobody should
reasonsbly using except in an emergency, so do not bother with trying to
maintain compatibility; it's a bit rediculous that we had 3 variations of
this to being with!

Signed-off-by: Sage Weil <sage@redhat.com>