pybind: add param option to argdesc in ceph_argparse.py
This patch adds an new parameter to the argdesc class, `param` that when
set to `true` will force the `--name=` syntax in the generated helpstr.

The `param=true` is also added to the `fomat` option of the
`ceph orchestrator device ls` command so the specified usage wont result
in a cli syntax error.

This patch also modifies the ceph client to remove any param=true
argdesc's from the srgs and deals with them non positionally.

Fixes: https://tracker.ceph.com/issues/42113
Signed-off-by: Matthew Oliver <moliver@suse.com>