mgr/MgrClient: fix open condition
Do not open a mgr daemon session if we are a client using a non-client
key to authenticate.  This is a bit of a corner case, but it can
happen when using the mon. key do bootstrap functions.

This aligns the MgrClient condition with the mgr/DaemonServer.cc
handle_report condition, which rejects non-daemon client connections
based on the connection entity type, not the auth identity.

Fixes: https://tracker.ceph.com/issues/42566
Signed-off-by: Sage Weil <sage@redhat.com>