global: disable THP for Ceph daemons
Ceph is known to suffer from memory fragmentation due to transparent
huge pages (THP). This is indicated by RSS usage above configured memory
targets. This is only observable when the distribution default for THP
is set to "always". In the upstream kernel, the default is "madvise".

For now, enabling THP is currently discouraged until selective use of
THP by Ceph is implemented via madvise. We will need to consider both
defaults for THP so madvise calls to both enable and disable THP will
need implemented.

All credit to Mark Nelson for doing the legwork identifying this issue.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>