PendingReleaseNotes: add telemetry to the very top
During cherry-picking we ended up with the notes for the features at the
very bottom. Instead, move them all to the very top, for the next
release.

We're doing this using a single patch because the PendingReleaseNotes
file was updated several times over the course of this patch set, and
reworking the whole thing would be annoying.

Signed-off-by: Joao Eduardo Luis <joao@suse.com>