ceph-daemon: refactor c-v activate/deactivate into systemd unit
For the activation, do the prestart step (c-v lvm activate) as part of the
main 'run' script.  This ensures that any setup (like dm-crypt) is done
every time the unit starts.  We could separate the activate step into a
separate PreStart hook in the systemd unit but don't see any benefit to
doing so.

For deactivation, add a poststop systemd hook and put that in the
'poststop' script.  This isn't enabled yet because c-v doesn't yet
implement deactivate.

Signed-off-by: Sage Weil <sage@redhat.com>