rbd-mirror: reset bootstrap state to replaying if local image promoted
The previous logic had a race in that it would never run if the local
image was force-promoted or if the remote image was demoted and the
local image was promoted before the rbd-mirror daemon could catch the
change of state.

This change resets the mirror peer state from syncing to replaying if
the local image is no longer connected to the remote image. This
prevents an unexpected (and corrupted) re-sync from the remote.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>
(cherry picked from commit 8c33391a6294f68b590f26fff190a34a2dc317eb)