cephadm: add alertmanager deployment feature
Deploy now accepts a daemon_type of alertmanager. Since alertmanager
is a cluster aware service, the monitoring metadata has been updated to
allow a daemon to use multiple ports. In addition, when config_json is
received, any 'key' prefixed by '_' is skipped when creating files in the
daemons etc directory. Keys that use the '_' prefix hold config data that
can be used elsewhere. In the case of the alertmanager a _peers parameter
is required which is used to add --cluster.peer=<ip>:<port> to the
container command to form the alertmanager cluster

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>