mon: elector: continue degrading scores on dead peer connections
We need to keep scheduling timer events, not doing one report_dead_connection()
and then stopping! So we add a new dead_ping() function that parallels the
ping_check(). Along with that, we need to clean up state when a peer is removed
or our rank changes to an existing one. Extend the notify_peer_* functions
to deal with that, and maintain the new dead_pinging set in parallel
with live_pinging.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>