common/cmdparse: implicit positional=false following n=N
There are lots of CephBool parameters this applies too, but they already
look like non-positional arguments (as far as I can tell, at least).

There is *one* other command:

COMMAND("pg dump_stuck "
	"name=stuckops,type=CephChoices,strings=inactive|unclean|stale|undersized|degraded,n=N,req=false " \
	"name=threshold,type=CephInt,req=false",

In this case previously you could do 'pg dump_stuck inactive 123' and the
parser could tell 123 was not part of the stuckops list because it wasn't
in the enum.  With this PR you now have to do

 ceph pg dump_stuck inactive --threshold 123

IMO this is an improvement as it more closely aligns with how other
parsers work and is less ambiguous.

Signed-off-by: Sage Weil <sage@redhat.com>